Modify Kuryr not to delete the default subnets
This patch fixes the bug that Kuryr deletes the default subnets
associated with the subnet pools for Kuryr specified in the config
file.

By default Kuryr creates the default subnets with the subnet pools. When
users delete the Docker endpoint, Kuryr deleted the default subnets.
However, the default subnets are used when users don't specify the CIDRs
for interfaces when they create Docker endpoints, so it's a better design
not to delete them.

Closes-Bug: #1497994

+++ kuryr/controllers.py
@@ -55,6 +55,23 @@ SUBNET_POOLS_V6 = [
 app.neutron.format = 'json'
 
 
+def _cache_default_subnetpool_ids(app):
+    """Chaches IDs of the default subnetpools as app.DEFAULT_POOL_IDS."""
+    if not hasattr(app, 'DEFAULT_POOL_IDS'):
+        default_subnetpool_ids = []
+        try:
+            subnetpool_names = SUBNET_POOLS_V4 + SUBNET_POOLS_V6
+            for subnetpool_name in subnetpool_names:
+                subnetpools = app.neutron.list_subnetpools(
+                    name=subnetpool_name)
+                for subnetpool in subnetpools['subnetpools']:
+                    default_subnetpool_ids.append(subnetpool['id'])
+        except n_exceptions.NeutronClientException as ex:
+            app.logger.error("Error happened during retrieving the default "
+                             "subnet pools.".format(ex))
+        app.DEFAULT_POOL_IDS = default_subnetpool_ids
+
+
 def _get_subnets_by_attrs(**attrs):
     subnets = app.neutron.list_subnets(**attrs)
     if len(subnets) > 1:
@@ -378,14 +395,22 @@ def network_driver_delete_endpoint():
             raise
 
         for subnet_id in concerned_subnet_ids:
-            # If the subnet to be deleted has any port, when some ports are
-            # referring to the subnets in other words, delete_subnet throws an
-            # exception, SubnetInUse that extends Conflict. This can happen
-            # when the multiple Docker endpoints are created with the same
-            # subnet CIDR and it's totally the normal case. So we'd just log
-            # that and continue to proceed.
             try:
-                app.neutron.delete_subnet(subnet_id)
+                subnet = app.neutron.show_subnet(subnet_id)
+                subnet = subnet['subnet']
+                pool_id = subnet.get('pool_id', None)
+
+                _cache_default_subnetpool_ids(app)
+
+                if pool_id not in app.DEFAULT_POOL_IDS:
+                    # If the subnet to be deleted has any port, when some ports
+                    # are referring to the subnets in other words,
+                    # delete_subnet throws an exception, SubnetInUse that
+                    # extends Conflict. This can happen when the multiple
+                    # Docker endpoints are created with the same subnet CIDR
+                    # and it's totally the normal case. So we'd just log that
+                    # and continue to proceed.
+                    app.neutron.delete_subnet(subnet_id)
             except n_exceptions.Conflict as ex:
                 app.logger.info("The subnet with ID {0} is still referred "
                                 "from other ports and it can't be deleted for "
